completed weather module
had to set it up as a class after
meeting with David.  Was unable to
get it to work as a module.  Couldn't
call multiple methods on that new
object.  Discussed with team and will
follow up with Phil, Max, and David.
Current focus is to get APIs working
in the application.